requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
django-filter
django-cors-headers